[enh] process log files with (small) constant memory usage even for very large log files
Original implementation dependent on zipped and unzipped size of log file
---
1. Download gzipped logfile into variable access_log_gz which can get large (e.g. 60MB)
2. Inflate into variable access_log (even larger, e.g. 300 MB)
3. Split access_log on newline and process each line
4. Do this in parallel on thread_num threads. Eight threads would use almost 3GB memory

New implementation independent of size of zipped or unzipped file
---
1. Download gzipped file to file system (returns file object)
2. Rewind file object
3. Create EnumerableInflater on file object process file in 1 MB chunks
4. Enumerate lines in the file and process each line
5. Do this in parallel on thread_num threads. Little extra memory usage

The EnumerableInflater In step 3 gets a 1 MB chunk of the binary (local) file,
unzips that chunk and passed the plain text to the next enumerator which scans
for newlines and return these lines to then next enumerator that is used to
iterate over all lines.

In this way the binary gzipped file is processed in a streaming fashion with
constant memory usage independent of size of (zipped and unzipped) log file.

Technically it is possible to stream chunks of binary data directly from S3, but in this
case the underlying implementation can't retry failed communication to S3. The
get_object call to download the whole file does implement retries and is therefore
a more robust mechanism to get a file from S3.